basic card skeleton finished
