+++ package.json
@@ -22,7 +22,8 @@
     "react-native": "^0.64.2",
     "react-native-web": "0.17.1",
     "react-redux": "^7.2.6",
-    "redux": "^4.1.1"
+    "redux": "^4.1.1",
+    "tailwind-rn": "^3.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
