--- server/util/geocoder.js
@@ -1,12 +1,0 @@
-const NodeGeocoder = require('node-geocoder');
-
-const options = {
-  provider: GEOCODER_PROVIDER,
-  httpAdapter: 'https',
-  apiKey: process.env.GEOCODER_API_KEY, // for Mapquest, OpenCage, Google Premier
-  formatter: null // 'gpx', 'string', ...
-};
-
-const geocoder = NodeGeocoder(options);
-
-module.exports = geocoder
